move cache-related Configuration code into Configuration+Cache.swift
